backend/requirements.txt
google-adk==1.27.1
nostr==0.0.2
cryptography==46.0.5
firebase-admin==7.2.0
python-dotenv==1.2.2
fastapi==0.135.1
uvicorn==0.41.0
websockets==15.0.1
